refactor: separate into files
