+++ src/graphql/artist/schema.ts
