+++ src/graphql/player/index.ts
@@ -1,2 +1,2 @@
-export { default as playerResolvers } from "./playerResolvers.js";
-export * from "./playerSchema.js";
+export { default as playerResolvers } from "./resolvers/index.js";
+export * from "./schema.js";
